backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-community
langchain-openai
langchain-anthropic
chromadb
python-dotenv
pydantic
pypdf
python-docx
sqlalchemy
